Revert "Merge pull request #907 from handokei/feat/#901-boarding-lock-sync"
This reverts commit 1b07e03b54f891def802bfde33030c2c1ad86c63, reversing
changes made to 926ff56fd9dfaaa50cb18bf0cd89ae9ed3c7755d.
