+++ backend/alarm-worker/src/index.ts
@@ -26,7 +26,7 @@ import {
 import { ackPending } from './pendingPushes';
 import { appendPositionPoint } from './positionSeries';
 import { appendAccelSample, isAccelSummary } from './accelSeries';
-import { deleteProgress, getProgress, putProgress, type TripProgress } from './progress';
+import { deleteProgress, getProgress, type TripProgress } from './progress';
 import { SeoulArrivalClient } from './seoul';
 import { runScheduled } from './scheduled';
 import {
@@ -356,158 +356,6 @@ app.post('/metrics/boarding-prompt', async (c) => {
   }
   return c.json({ ok: true });
 });
-
-/**
- * Seam E (#901) — 지상 BoardingLock 정정 채널.
- *
- * 클라가 좋은 GPS fix(accuracy ≤ 50m)로 현재 정거장을 확정했을 때 호출. backend는 관측 역이
- * 진행 방향 ±1 hop 이내면 waypoints를 advance해 lock의 currentWaypoint를 사용자 실제 위치와
- * 정렬한다. cron의 Seoul 도착 폴링이 stale일 때(터널/긴 지연) silent push 누락 회귀(#622) 흡수.
- *
- * 지하 구간은 정의상 GPS 부재 → 클라가 호출하지 않는다. 호출 트리거는 클라 책임:
- *   1) accuracy ≤ 50m + 새 currentStation 확정 (debounce 5s, useBoardingLockSync)
- *   2) 지하→지상 경계 (subsurface=false 전환, Seam G barometer)
- *   3) trip 등록 직후 1회
- *
- * Body: { token, observedStationName, observedAtMs, accuracy, subsurface? }
- * Response 200:
- *   { ok, advanced, currentWaypoint, nextStation }
- *   - advanced: 이번 sync로 waypoints가 shift됐는지 (1+ hop)
- *   - currentWaypoint: 정정 후 first waypoint stationName (없으면 null — destination 도착)
- *   - nextStation: 정정 후 first waypoint = 다음 알람 대상 (currentWaypoint와 동일, 의미상 alias)
- * Response 404: { error: 'trip_not_found' } — 클라는 다음 fix에서 자연 retry
- *
- * Trip 부재 시 lock 재생성 책임은 본 endpoint가 지지 않음 — 클라가 useApnsTripRegistration으로
- * POST /trips를 호출하면 같은 경로로 lock이 들어온다 (분리된 lock store가 없는 현 backend 구조).
- */
-app.post('/boarding-lock/sync', async (c) => {
-  let body: unknown;
-  try {
-    body = await c.req.json();
-  } catch {
-    return c.json({ error: 'invalid_json' }, 400);
-  }
-  const payload = validateBoardingLockSync(body);
-  if (!payload) return c.json({ error: 'invalid_payload' }, 400);
-
-  const existing = await getTrip(c.env.TRIPS, payload.token);
-  if (!existing) return c.json({ error: 'trip_not_found' }, 404);
-
-  const now = Date.now();
-  const advance = computeLockSyncAdvance(existing.waypoints, payload.observedStationName);
-
-  let working: Trip = existing;
-  if (advance.shiftedCount > 0) {
-    const remaining = working.waypoints.slice(advance.shiftedCount);
-    working = {
-      ...working,
-      waypoints: remaining,
-      // 새 waypoint의 첫 push를 보장하기 위해 baseline reset (advanceBoardingLockWaypoint와 동형).
-      lastTrackedArrivalEpoch: undefined,
-      lastLaPushEpoch: undefined,
-    };
-    // progress KV mirror — POST /trips re-register 시 같은 trainCode면 shift 진행분이 보존되도록.
-    await maybeMirrorLockSyncProgress(c.env.TRIPS, working, advance.shiftedCount);
-  }
-
-  // lock TTL refresh — 사용자가 지상에서 lock을 활성 유지 중임을 confirm.
-  if (working.boardingLock) {
-    working = {
-      ...working,
-      boardingLock: {
-        ...working.boardingLock,
-        expiresAt: Math.max(working.boardingLock.expiresAt, now + LOCK_TTL_REFRESH_MS),
-      },
-    };
-  }
-
-  await putTrip(c.env.TRIPS, working);
-
-  const head = working.waypoints[0];
-  return c.json({
-    ok: true,
-    advanced: advance.shiftedCount > 0,
-    currentWaypoint: head ? head.stationName : null,
-    nextStation: head ? head.stationName : null,
-  });
-});
-
-/** Seam E 정정으로 lock TTL을 연장하는 길이. cron 주기 60s × 30 cycles 마진. */
-export const LOCK_TTL_REFRESH_MS = 30 * 60 * 1000;
-
-interface BoardingLockSyncPayload {
-  token: string;
-  observedStationName: string;
-  observedAtMs: number;
-  accuracy: number;
-  subsurface?: boolean;
-}
-
-export function validateBoardingLockSync(input: unknown): BoardingLockSyncPayload | null {
-  if (!input || typeof input !== 'object') return null;
-  const obj = input as Record<string, unknown>;
-  if (typeof obj.token !== 'string' || obj.token.length === 0) return null;
-  if (typeof obj.observedStationName !== 'string' || obj.observedStationName.length === 0) {
-    return null;
-  }
-  if (typeof obj.observedAtMs !== 'number' || !Number.isFinite(obj.observedAtMs)) return null;
-  if (typeof obj.accuracy !== 'number' || !Number.isFinite(obj.accuracy) || obj.accuracy < 0) {
-    return null;
-  }
-  const result: BoardingLockSyncPayload = {
-    token: obj.token,
-    observedStationName: obj.observedStationName,
-    observedAtMs: obj.observedAtMs,
-    accuracy: obj.accuracy,
-  };
-  if (typeof obj.subsurface === 'boolean') result.subsurface = obj.subsurface;
-  return result;
-}
-
-/**
- * Seam E 진행 판단 — 관측 역이 waypoints 시퀀스 안 어디인지 찾고 shift 개수를 산출.
- *
- * 정책:
- *   - waypoints[0] 일치: 현재 다음 hop 도달 → 1 hop advance
- *   - waypoints[1] 일치: 1 hop 앞서감 (cron이 한 사이클 늦었음) → 2 hop catch-up advance
- *   - waypoints[k≥2] 일치: k hop catch-up advance (긴 음영 후 재진입 케이스)
- *   - 미일치: 사용자가 진행 방향 뒤에 있거나 다른 트립 → no-op (lock 보존)
- *
- * "사용자 뒤 1 hop은 grace 1 cycle 후 advance"는 Seam E가 아닌 cron의 자연 추적이 담당 —
- * 본 endpoint는 GPS-확신 신호만 받아 진행 정정에 집중 (역방향 advance 안 함).
- */
-export function computeLockSyncAdvance(
-  waypoints: Trip['waypoints'],
-  observedStationName: string,
-): { shiftedCount: number } {
-  const idx = waypoints.findIndex((w) => w.stationName === observedStationName);
-  if (idx < 0) return { shiftedCount: 0 };
-  return { shiftedCount: idx + 1 };
-}
-
-/**
- * Seam E의 advance도 progress KV에 mirror — POST /trips 재등록 race에서 shift 진행분 보존.
- * lock(trainCode) 없는 trip은 progress 자체가 의미 없어 no-op (scheduled.ts mirrorProgress와 동형).
- */
-async function maybeMirrorLockSyncProgress(
-  kv: KVNamespace,
-  trip: Trip,
-  shiftedDelta: number,
-): Promise<void> {
-  const trainCode = trip.boardingLock?.trainCode;
-  if (!trainCode) return;
-  const existing = await getProgress(kv, trip.token);
-  const prevShifted = existing?.trainCode === trainCode ? existing.shiftedCount : 0;
-  const next: TripProgress = {
-    trainCode,
-    shiftedCount: prevShifted + shiftedDelta,
-    lastTrackedArrivalEpoch: trip.lastTrackedArrivalEpoch,
-    lastLaPushEpoch: trip.lastLaPushEpoch,
-    consecutiveEtaMissing: trip.consecutiveEtaMissing,
-  };
-  const ttlSec = Math.max(60, Math.floor((trip.expiresAt - Date.now()) / 1000));
-  await putProgress(kv, trip.token, next, ttlSec);
-}
 
 interface PushAckPayload {
   pushId: string;
